requirements.txt
python-docx
openpyxl
docx2pdf
Pillow
